+++ tests/test_selectors.py
@@ -110,7 +110,7 @@ def test_with_memory_priority_handling():
 def test_tick_add_with_current_child():
     console.banner('Tick-Add with Current Child')
     assert_banner()
-    root = py_trees.composites.Selector(name="Selector")
+    root = py_trees.composites.Selector(name="Selector", memory=False)
     root.tick_once()
     child = py_trees.behaviours.Failure(name="Failure")
     print(py_trees.display.unicode_tree(root, show_status=True))
@@ -123,7 +123,7 @@ def test_tick_add_with_current_child():
 def test_add_tick_add_with_current_child():
     console.banner('Add-Tick-Add with Current Child')
     assert_banner()
-    root = py_trees.composites.Selector(name="Selector")
+    root = py_trees.composites.Selector(name="Selector", memory=False)
     run1 = py_trees.behaviours.Running("Run1")
     run2 = py_trees.behaviours.Running("Run2")
     root.add_child(run1)
@@ -140,7 +140,7 @@ def test_add_tick_add_with_current_child():
 def test_add_tick_insert_with_current_child():
     console.banner('Add-Tick-Insert with Current Child')
     assert_banner()
-    root = py_trees.composites.Selector(name="Selector")
+    root = py_trees.composites.Selector(name="Selector", memory=False)
     run1 = py_trees.behaviours.Running("Run1")
     run2 = py_trees.behaviours.Running("Run2")
     root.add_child(run1)
@@ -157,7 +157,7 @@ def test_add_tick_insert_with_current_child():
 def test_add_tick_remove_with_current_child():
     console.banner('Add-Tick-Remove with Current Child')
     assert_banner()
-    root = py_trees.composites.Selector(name="Selector")
+    root = py_trees.composites.Selector(name="Selector", memory=False)
     child = py_trees.behaviours.Success(name="Success")
     root.add_child(child)
     root.tick_once()